Rimosso Magdwick e aggiunti parametri per i PID e le dimensioni delle ruote e della loro distanza

+++ BuggyROS446/Core/Inc/ros/node_handle.h
@@ -587,7 +587,7 @@ public:
   {
     if (requestParam(name, timeout))
     {
-      if (length == req_param_resp.floats_length)
+      if (uint32_t(length) == req_param_resp.floats_length)
       {
         //copy it over
         for (int i = 0; i < length; i++)
